Isolating path class in a separate script
