+++ reg_train.py
@@ -8,7 +8,7 @@
 import pandas as pd
 import os
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
-from base_model import PathConfig
+from paths import PathConfig
 from model_lib import ModelConfig
 
 def main():
